requirements.txt
boto3>=1.28.0
bcrypt>=4.0.0
PyJWT>=2.8.0